Floating-point math should have no place in an ircd
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@12442 e03df62e-2008-0410-955e-edbf42e46eb7

+++ include/ctables.h
@@ -108,11 +108,11 @@ class CoreExport Command : public ServiceProvider
 
 	/** used by /stats m
 	 */
-	long double use_count;
+	long use_count;
 
 	/** used by /stats m
 	 */
-	long double total_bytes;
+	long total_bytes;
 
 	/** True if the command is disabled to non-opers
 	 */
